[REFAC: #6] - Extraindo mais componentes.
close #6
